requirements.txt
alembic==1.4.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
python-dotenv==0.12.0
python-jose==3.1.0
SQLAlchemy==1.3.15
